[libmseed] Fix a possible collision of static variables between Python modules (#55)
This PR fixes a problem that can occur with Python modules using the SeisComP Python API along with 3rd-party Python packages using libmseed. If both SeisComP and the 3rd-party package configure their logging routines using libmseed `ms_loginit()`, one will override the settings of the other, resulting in dangling pointers. As soon as one of the functions pointed to by `log_print`, or `diag_print` is called e.g. in [1], a segfault results.

By renaming the static libmseed variable `gMSLogParam` in SeisComP, this collision is avoided.

[1] https://github.com/SeisComP/seiscomp/blob/ba124aedd9c0a9ad9708675934891d0bcbb5f0ac/src/system/libs/3rd-party/libmseed/logging.c#L295